style(cpp): drop gratuitous static_cast<size_t> on container subscripts (#1506)
Remove 20 `x[static_cast<size_t>(...)]` casts where the index is already a signed type (`i_t` / `cuopt_int_t` / `int64_t`). The build uses `-Werror` but not `-Wsign-conversion`/`-Wconversion` and has no clang-tidy, so these casts suppressed no warning and only added noise inconsistent with the surrounding bare-signed-indexing style.

Legitimate casts are kept: container sizing from signed subtraction, the `size_t`->`i_t` narrowing of `.size()`, and the hoisted `col_u` variable.

Verified with clang-format 20.1.8 and a full `./build.sh libcuopt` (-Werror) pass.

See #1505 for the style guidance change.

Authors:
  - Miles Lubin (https://github.com/mlubin)

Approvers:
  - Nicolas L. Guidotti (https://github.com/nguidotti)

URL: https://github.com/NVIDIA/cuopt/pull/1506

+++ cpp/src/grpc/server/grpc_service_impl.cpp
@@ -884,7 +884,7 @@ class CuOptRemoteServiceImpl final : public cuopt::remote::CuOptRemoteService::S
     if (max_count > 0 && count > max_count) { count = max_count; }
 
     for (int64_t i = 0; i < count; ++i) {
-      const auto& inc = incumbents[static_cast<size_t>(from_index + i)];
+      const auto& inc = incumbents[from_index + i];
       auto* out       = response->add_incumbents();
       out->set_index(from_index + i);
       out->set_objective(inc.objective);
